Spawn object sprites asynchronously
